对密码修改那快进行了功能优化,添加验证码修改,然后美化了管理端的这个用户管理界面显示

+++ interview-backend/src/main/java/com/interview/service/AdminUserService.java
@@ -122,6 +122,9 @@ public class AdminUserService {
     }
 
     public void delete(Long id) {
+        if (StpUtil.getLoginIdAsLong() == id) {
+            throw new BizException(40000, "不能删除当前登录账号");
+        }
         userMapper.deleteById(id);
         StpUtil.kickout(id);
     }
